agent 03/requirements_reviewer.txt
fastapi
uvicorn
python-multipart
python-dotenv
# AI Providers
openai
anthropic
google-generativeai
# File Parsers
pdfplumber
pandas
openpyxl
python-docx